Upgrade phan and unbreak master
Follow-up: I606d6205ee22e48ca6df04c14954ed7cefcadd64 which left an issue unfixed

Also fixes a major bug in SpecialImageRating which caused the 'categorylinks'
table to never be used in SELECTs.
And also a bug where a locale-formatted number was stored in cache and passed
around, and notably it was fed to VoteStars::displayStars, which expects an
integer and will probably fail unexpectedly if given a localised number.

+++ includes/api/ApiImageRating.php
@@ -43,6 +43,7 @@ class ApiImageRating extends ApiBase {
 
 		// Top level
 		$this->getResult()->addValue( null, $this->getModuleName(),
+			// @phan-suppress-next-line PhanTypeMismatchArgument T240141
 			[ 'result' => self::addImageCategory( $pageId, $params['categories'] ) ]
 		);
 
@@ -64,7 +65,7 @@ class ApiImageRating extends ApiBase {
 		$wp = WikiPage::factory( $imagePage );
 
 		// Check if it's been edited in last 2 seconds: want to delay the edit
-		$timeSinceEdited = wfTimestamp( TS_MW, 0 ) - $wp->getTimestamp();
+		$timeSinceEdited = (int)wfTimestamp( TS_MW, 0 ) - (int)$wp->getTimestamp();
 		if ( $timeSinceEdited <= 2 ) {
 			return 'busy';
 		}
